+++ flask_jinjahelpers/__init__.py
@@ -1,7 +1,5 @@
-from copy import copy
 from flask import request, url_for
 from jinja2 import ChoiceLoader, PackageLoader
-from werkzeug.datastructures import MultiDict
 
 
 def init_jinja_env(env):
@@ -26,24 +24,6 @@ def init_jinja_env(env):
     return env
 
 
-def qp_url_for(endpoint, **kwargs):
-    """
-    THIS FUNCTION HAS BEEN DEPRECATED:
-
-    Returns the url for an endpoint but preserves all request query parameters
-
-    kwargs can be used for overriding specific query parameters
-
-    :param endpoint: endpoint to return the url for, eg. user.index
-    :param kwargs: dict containing query parameter names as keys
-    """
-    data = dict(MultiDict(request.args).lists())
-
-    for key, value in kwargs.items():
-        data[key] = value
-    return url_for(endpoint, **data)
-
-
 def url_for_current(callback=None, **kwargs):
     """
     Returns the current url endpoint with all query parameters but replaces
@@ -62,44 +42,40 @@ def url_for_current(callback=None, **kwargs):
     return url_for(request.endpoint, **data)
 
 
-def header_sort_url(
-    sort_by, sorted_fields=[], max_sorted_fields=2, **kwargs
-):
-    if not sort_by:
-        return url_for_current(**kwargs)
-
+def inverse_sorted_field(sort_by, sorted_fields):
     if isinstance(sorted_fields, basestring):
         sorted_fields = [sorted_fields]
+    elif sorted_fields is None:
+        sorted_fields = []
 
-    sorted_fields_copy = copy(sorted_fields)
-    field_names = []
-
-    for index, sorted_field in enumerate(sorted_fields):
-        if (not sorted_field or
-                (len(sorted_field) == 1 and sorted_field[0] == '-')):
-            del sorted_fields_copy[index]
-            continue
-
-        if sorted_field[0] == '-':
-            field_names.append(sorted_field[1:])
-        else:
-            field_names.append(sorted_field)
+    sorted_fields = [s for s in sorted_fields if s.strip('-')]
+    field_names = [s.strip('-') for s in sorted_fields]
 
     try:
         field_index = field_names.index(sort_by)
     except ValueError:
-        sorted_fields_copy.insert(0, sort_by)
+        sorted_fields.insert(0, sort_by)
     else:
-        old_sort_by = sorted_fields_copy[field_index]
-        del sorted_fields_copy[field_index]
+        old_sort_by = sorted_fields[field_index]
+        del sorted_fields[field_index]
         if old_sort_by[0] == '-':
-            sorted_fields_copy.insert(0, sort_by)
+            sorted_fields.insert(0, sort_by)
         else:
-            sorted_fields_copy.insert(0, '-%s' % sort_by)
+            sorted_fields.insert(0, '-%s' % sort_by)
+    return sorted_fields
 
-    sorted_fields_copy = sorted_fields_copy[0:max_sorted_fields]
-    if sorted_fields_copy:
-        kwargs['sort'] = sorted_fields_copy
+
+def header_sort_url(
+    sort_by, sorted_fields=None, max_sorted_fields=1, **kwargs
+):
+    if not sort_by:
+        return url_for_current(**kwargs)
+
+    sorted_fields = inverse_sorted_field(sort_by, sorted_fields)
+    sorted_fields = sorted_fields[0:max_sorted_fields]
+
+    if sorted_fields:
+        kwargs['sort'] = sorted_fields
     return url_for_current(**kwargs)